+++ samples/showcase/build.gradle.kts
@@ -1,0 +1,19 @@
+import org.jetbrains.compose.ComposeBuildConfig
+
+plugins {
+    jewel
+    `jewel-publish`
+    `jewel-check-public-api`
+    alias(libs.plugins.composeDesktop)
+    alias(libs.plugins.compose.compiler)
+}
+
+private val composeVersion
+    get() = ComposeBuildConfig.composeVersion
+
+dependencies {
+    implementation(project(":foundation"))
+    implementation(project(":ui"))
+    testImplementation(compose.desktop.uiTestJUnit4)
+    testImplementation(compose.desktop.currentOs) { exclude(group = "org.jetbrains.compose.material") }
+}
