+++ settings.gradle.kts
@@ -44,6 +44,7 @@ include(
     ":markdown:int-ui-standalone-styling",
     ":markdown:ide-laf-bridge-styling",
     ":samples:ide-plugin",
+    ":samples:showcase",
     ":samples:standalone",
     ":ui",
     ":ui-test",
